Fix score-field overflow and map-bounds for segment-only maps
Two more code-audit fixes:

- H7: the playerScores packet serialized kills/assists/deaths as $uint8,
  so they silently wrapped at 256 in a long match (300 kills -> 44).
  Widened those three fields to $uint16 (damage stays $uint32). The encode
  helper forwards plain numbers, so the packet definition drives both
  directions consistently. Regression test round-trips a 300/512/1000 score.
- M2: map bounds were computed only from spawn/wall tiles, ignoring
  wall_segments — a segment-only map clamped ships inside its own walls, and
  a tile-less map produced inverted bounds (min=+Inf, max=-Inf) that broke
  applyMapBounds. Now both segment endpoints feed the bounds, with a default
  box fallback when a map has no tiles and no segments. Tests cover a
  segment-only map (bounds enclose segments) and an empty map (default box).

109 tests pass; game build + lint clean.

+++ packages/game/src/networking/packets.ts
@@ -133,9 +133,9 @@ export const packetManager = new PacketManager({
 
     playerScores: new Packet({
         playerId: $string(CONNECTION_ID_LENGTH),
-        kills: $uint8,
-        assists: $uint8,
-        deaths: $uint8,
+        kills: $uint16,
+        assists: $uint16,
+        deaths: $uint16,
         damage: $uint32,
     }),
